Feat & refact: add open and close photo modal functions and restructure modules
