+++ js/main.js
@@ -1,8 +1,6 @@
 import { createPhotosDescriptions } from './data-generators/photos-descriptions.js';
-import { drawUsersPhotos } from './draw-users-photos.js';
-import { onPhotoClick } from './open-photo-modal.js';
+import { renderThumbnails } from './render-thumbnails.js';
 
 const usersPhotos = createPhotosDescriptions();
 
-drawUsersPhotos(usersPhotos);
-onPhotoClick(usersPhotos);
+renderThumbnails(usersPhotos);
