+++ js/utils/keyboard-utils.js
@@ -1,0 +1,5 @@
+const isEscapeKey = (evt) => evt.key === 'Escape';
+
+const isEnterKey = (evt) => evt.key === 'Enter';
+
+export { isEscapeKey, isEnterKey }
